+++ actix-redis/src/lib.rs
@@ -1,5 +1,6 @@
 //! Redis integration for `actix`.
 
+#![forbid(unsafe_code)]
 #![deny(rust_2018_idioms, nonstandard_style)]
 #![warn(future_incompatible)]