+++ actix-web-httpauth/CHANGES.md
@@ -1,6 +1,13 @@
 # Changes
 
-## Unreleased - 2021-xx-xx
+## Unreleased - 2022-xx-xx
+
+
+## 0.7.0 - 2022-07-19
+- Auth validator functions now need to return `(Error, ServiceRequest)` in error cases. [#260]
+- Minimum supported Rust version (MSRV) is now 1.57 due to transitive `time` dependency.
+
+[#260]: https://github.com/actix/actix-extras/pull/260
 
 
 ## 0.6.0 - 2022-03-01